move status indicator to separate component and move colors.json to colors.ts
