--- MeowdoptMeApp/assets/colors.json
@@ -1,6 +1,0 @@
-{
-  "backgroundColor": "#CBFFFAFF",
-  "lightAccentColor": "#FFFFFFFF",
-  "darkAccentColor": "#78ECDFFF",
-  "strongAccentColor": "#199789FF"
-}
