+++ MeowdoptMeApp/assets/colors.ts
@@ -1,0 +1,8 @@
+const colorPalette = {
+  backgroundColor: '#CBFFFAFF',
+  lightAccentColor: '#FFFFFFFF',
+  darkAccentColor: '#78ECDFFF',
+  strongAccentColor: '#199789FF',
+};
+
+export default colorPalette;
